+++ src/utils/cssRuleIsKeyframe.js
@@ -1,0 +1,13 @@
+/**
+ * Check if a rule is a keyframe one
+ *
+ * @param {Rule} rule - postcss rule node
+ * @return {boolean} True if the rule is a keyframe one
+ */
+export default function (rule) {
+  const { parent } = rule
+  return (
+    parent.type === "atrule"
+    && parent.name === "keyframes"
+  )
+}