+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fixed: `max-line-length` options validation.
 - Fixed: `function-calc-no-unspaced-operator` accepts newlines.
 - Fixed: `block-closing-brace-newline-after` accepts single-line comments immediately after the closing brace.
+- Fixed: `selector-no-id` now ignores keyframe selectors.
 
 # 5.2.1